requirements.txt
mysql-connector-python==2.2.9
reportlab==4.0.0
customtkinter==5.1.2
ttkbootstrap==1.10.1
